+++ wemake_python_styleguide/violations/oop.py
@@ -94,7 +94,6 @@ class BuiltinSubclassViolation(ASTViolation):
 
     error_template = 'Found subclassing a builtin: {0}'
     code = 600
-    previous_codes = {426}
 
 
 @final
@@ -148,7 +147,6 @@ class ShadowedClassAttributeViolation(ASTViolation):
 
     error_template = 'Found shadowed class attribute: {0}'
     code = 601
-    previous_codes = {427}
 
 
 @final
@@ -170,7 +168,6 @@ class StaticMethodViolation(ASTViolation):
 
     error_template = 'Found using `@staticmethod`'
     code = 602
-    previous_codes = {433}
 
 
 @final
@@ -201,7 +198,6 @@ class BadMagicMethodViolation(ASTViolation):
 
     error_template = 'Found using restricted magic method: {0}'
     code = 603
-    previous_codes = {434}
 
 
 @final
@@ -238,7 +234,6 @@ class WrongClassBodyContentViolation(ASTViolation):
 
     error_template = 'Found incorrect node inside `class` body'
     code = 604
-    previous_codes = {452}
 
 
 @final
@@ -274,7 +269,6 @@ class MethodWithoutArgumentsViolation(ASTViolation):
 
     error_template = 'Found method without arguments: {0}'
     code = 605
-    previous_codes = {453}
 
 
 @final
@@ -312,7 +306,6 @@ class WrongBaseClassViolation(ASTViolation):
 
     error_template = 'Found incorrect base class'
     code = 606
-    previous_codes = {454}
 
 
 @final
@@ -360,7 +353,6 @@ class WrongSlotsViolation(ASTViolation):
 
     error_template = 'Found incorrect `__slots__` syntax'
     code = 607
-    previous_codes = {455}
 
 
 @final
@@ -392,7 +384,6 @@ class WrongSuperCallViolation(ASTViolation):
 
     error_template = 'Found incorrect `super()` call: {0}'
     code = 608
-    previous_codes = {456}
 
 
 @final
@@ -435,7 +426,6 @@ class DirectMagicAttributeAccessViolation(ASTViolation):
 
     error_template = 'Found direct magic attribute usage: {0}'
     code = 609
-    previous_codes = {462}
 
 
 @final
@@ -522,7 +512,6 @@ class YieldMagicMethodViolation(ASTViolation):
 
     error_template = 'Found forbidden `yield` magic method usage: {0}'
     code = 611
-    previous_codes = {439, 435}
 
 
 @final